fix: make generateid available for others to use
