+++ internal/command/command_plugin.go
@@ -14,10 +14,10 @@ import (
 	mpi "github.com/nginx/agent/v3/api/grpc/mpi/v1"
 	"github.com/nginx/agent/v3/internal/bus"
 	"github.com/nginx/agent/v3/internal/config"
-	"github.com/nginx/agent/v3/internal/datasource/proto"
 	"github.com/nginx/agent/v3/internal/grpc"
 	"github.com/nginx/agent/v3/internal/logger"
 	pkgConfig "github.com/nginx/agent/v3/pkg/config"
+	"github.com/nginx/agent/v3/pkg/id"
 )
 
 var _ bus.Plugin = (*CommandPlugin)(nil)
@@ -275,7 +275,7 @@ func (cp *CommandPlugin) createDataPlaneResponse(correlationID string, status mp
 ) *mpi.DataPlaneResponse {
 	return &mpi.DataPlaneResponse{
 		MessageMeta: &mpi.MessageMeta{
-			MessageId:     proto.GenerateMessageID(),
+			MessageId:     id.GenerateMessageID(),
 			CorrelationId: correlationID,
 			Timestamp:     timestamppb.Now(),
 		},
