+++ internal/command/command_plugin_test.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	pkg "github.com/nginx/agent/v3/pkg/config"
+	"github.com/nginx/agent/v3/pkg/id"
 
-	"github.com/nginx/agent/v3/internal/datasource/proto"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/nginx/agent/v3/internal/bus/busfakes"
@@ -340,7 +340,7 @@ func TestMonitorSubscribeChannel(t *testing.T) {
 func Test_createDataPlaneResponse(t *testing.T) {
 	expected := &mpi.DataPlaneResponse{
 		MessageMeta: &mpi.MessageMeta{
-			MessageId:     proto.GenerateMessageID(),
+			MessageId:     id.GenerateMessageID(),
 			CorrelationId: "dfsbhj6-bc92-30c1-a9c9-85591422068e",
 			Timestamp:     timestamppb.Now(),
 		},
