+++ internal/file/file_manager_service.go
@@ -16,7 +16,6 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/nginx/agent/v3/internal/datasource/proto"
 	"github.com/nginx/agent/v3/internal/model"
 
 	"github.com/cenkalti/backoff/v4"
@@ -26,6 +25,7 @@ import (
 	"github.com/nginx/agent/v3/internal/grpc"
 	"github.com/nginx/agent/v3/internal/logger"
 	"github.com/nginx/agent/v3/pkg/files"
+	"github.com/nginx/agent/v3/pkg/id"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	backoffHelpers "github.com/nginx/agent/v3/internal/backoff"
@@ -104,7 +104,7 @@ func (fms *FileManagerService) UpdateOverview(
 
 	request := &mpi.UpdateOverviewRequest{
 		MessageMeta: &mpi.MessageMeta{
-			MessageId:     proto.GenerateMessageID(),
+			MessageId:     id.GenerateMessageID(),
 			CorrelationId: correlationID,
 			Timestamp:     timestamppb.Now(),
 		},
@@ -224,7 +224,7 @@ func (fms *FileManagerService) UpdateFile(
 			Contents: contents,
 		},
 		MessageMeta: &mpi.MessageMeta{
-			MessageId:     proto.GenerateMessageID(),
+			MessageId:     id.GenerateMessageID(),
 			CorrelationId: correlationID,
 			Timestamp:     timestamppb.Now(),
 		},
@@ -382,7 +382,7 @@ func (fms *FileManagerService) fileUpdate(ctx context.Context, file *mpi.File) e
 	getFile := func() (*mpi.GetFileResponse, error) {
 		return fms.fileServiceClient.GetFile(ctx, &mpi.GetFileRequest{
 			MessageMeta: &mpi.MessageMeta{
-				MessageId:     proto.GenerateMessageID(),
+				MessageId:     id.GenerateMessageID(),
 				CorrelationId: logger.GetCorrelationID(ctx),
 				Timestamp:     timestamppb.Now(),
 			},
