+++ pkg/id/message.go
@@ -3,17 +3,13 @@
 // This source code is licensed under the Apache License, Version 2.0 license found in the
 // LICENSE file in the root directory of this source tree.
 
-package proto
+package id
 
 import (
 	"log/slog"
 	"time"
 
-	mpi "github.com/nginx/agent/v3/api/grpc/mpi/v1"
-	"google.golang.org/protobuf/types/known/timestamppb"
-
 	"github.com/google/uuid"
-	agentUuid "github.com/nginx/agent/v3/pkg/uuid"
 )
 
 // UUIDGenerator defines a function type for generating UUIDs.
@@ -27,26 +23,8 @@ func GenerateMessageID() string {
 	uuidv7, err := defaultUUIDGenerator()
 	if err != nil {
 		slog.Debug("Issue generating uuidv7, using sha256 and timestamp instead", "error", err)
-		return agentUuid.Generate("%s", time.Now().String())
+		return Generate("%s", time.Now().String())
 	}
 
 	return uuidv7.String()
-}
-
-func CreateDataPlaneResponse(correlationID string, status mpi.CommandResponse_CommandStatus,
-	message, instanceID, err string,
-) *mpi.DataPlaneResponse {
-	return &mpi.DataPlaneResponse{
-		MessageMeta: &mpi.MessageMeta{
-			MessageId:     GenerateMessageID(),
-			CorrelationId: correlationID,
-			Timestamp:     timestamppb.Now(),
-		},
-		CommandResponse: &mpi.CommandResponse{
-			Status:  status,
-			Message: message,
-			Error:   err,
-		},
-		InstanceId: instanceID,
-	}
 }
