+++ pkg/id/uuid_test.go
@@ -3,7 +3,7 @@
 // This source code is licensed under the Apache License, Version 2.0 license found in the
 // LICENSE file in the root directory of this source tree.
 
-package uuid
+package id
 
 import (
 	"testing"
