+++ test/mock/grpc/mock_management_file_service.go
@@ -14,7 +14,7 @@ import (
 	"strconv"
 
 	"github.com/nginx/agent/v3/api/grpc/mpi/v1"
-	"github.com/nginx/agent/v3/internal/datasource/proto"
+	"github.com/nginx/agent/v3/pkg/id"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/encoding/protojson"
@@ -76,7 +76,7 @@ func (mgs *FileService) UpdateOverview(
 
 	configUploadRequest := &v1.ManagementPlaneRequest{
 		MessageMeta: &v1.MessageMeta{
-			MessageId:     proto.GenerateMessageID(),
+			MessageId:     id.GenerateMessageID(),
 			CorrelationId: request.GetMessageMeta().GetCorrelationId(),
 			Timestamp:     timestamppb.Now(),
 		},
